refactoring testrunner - some functions moved to abstract class
